Fix some bugs in LowCodeGen
I discovered that in Robert's code generates sv wrappers for
parameters---something that probably shouldn't be the case (?).
I removed this for now, but in any case, we should clarify:

(1) What are the semantics of parameter values
(2) When can we take a pointer/reference to something
(3) Can we take a pointer/reference to a parameter

I also changed the TR state monad to also capture procedure itself, to
avoid having to thread it through everywhere as a parameter.